2 commit

+++ conftest.py
@@ -4,7 +4,7 @@ from urls import Urls
 from data import generate_random_string
 
 @pytest.fixture
-def register_new_courier():
+def courier_data():
     login = generate_random_string(10)
     password = generate_random_string(10)
     first_name = generate_random_string(10)
@@ -15,15 +15,14 @@ def register_new_courier():
         'firstName': first_name
     }
 
-    response = requests.post(Urls.url_courier, data=payload)
+    yield payload
 
     login_payload = {
         'login': login,
         'password': password
     }
     login_response = requests.post(Urls.url_courier_login, data=login_payload)
-    login_data = login_response.json()
-    courier_id = login_data.get('id')
-
-    yield response, payload, courier_id
-    delete_response = requests.delete(f"{Urls.url_courier}/{courier_id}")
+    if login_response.status_code == 200:
+        courier_id = login_response.json().get('id')
+        if courier_id:
+            requests.delete(f"{Urls.url_courier}/{courier_id}")
